feat(Writing):
- Logic to control when to send POST /views/slug by using localStorage with TTL for each slug.
This plays nice with SWR with multiple tabs we can have the same value. So now when the user browse a post we add it to localStorage with a 5 min TTL
( If the user navigate to the post in this 5 minute window it will be counted as 1 view )

Not the best approach but better than un-realistic numbers, since next will call the api every time you navigate.
We also save bandwith with planetscale

- Planetscale development database is shared on my local machine and Vercel instance, the production one is added to the server.

- Updated environment example file